% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.R
\name{div}
\alias{div}
\title{Funcion division}
\usage{
div(a, b)
}
\arguments{
\item{a}{primer parametro}

\item{b}{segundo parametro}
}
\value{
la division
}
\description{
Funcion division
}
\examples{
div(1,1)

}
\seealso{
\link{mult}
}
