% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacion.R
\name{mult}
\alias{mult}
\title{Funcion multiplicacion}
\usage{
mult(a, b)
}
\arguments{
\item{a}{primer parametro}

\item{b}{segundo parametro}
}
\value{
la multiplicacion
}
\description{
Funcion multiplicacion
}
\examples{
mult(1,1)

}
\seealso{
\link{div}
}
