% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Funcion resta}
\usage{
resta(a, b)
}
\arguments{
\item{a}{primer parametro}

\item{b}{segundo parametro}
}
\value{
la resta
}
\description{
Funcion resta
}
\examples{
resta(1,1)

}
\seealso{
\link{suma}
}
