% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Funcion suma}
\usage{
suma(a, b)
}
\arguments{
\item{a}{primer parametro}

\item{b}{segundo parametro}
}
\value{
la suma
}
\description{
Funcion suma
}
\examples{
suma(1,1)

}
\seealso{
\link{resta}
}
